Fixed bug in parse_type<float>
git-svn-id: http://hopmod.googlecode.com/svn/trunk@417 d7818858-e1b4-11dd-a738-5fe2e3cfbcc5

+++ src/fpsgame/hopmod/libcubescript/include/cubescript/lexcast.hpp
@@ -84,7 +84,7 @@ template<> inline
 float parse_type<float>(const std::string & str)
 {
     if(!str.length()) return 0.0;
-    int val; 
+    float val;
     parse_type(str,val);
     return val;
 }
